Separate useAlertsPageData.js
