+++ src/elements/AlertsTable/AlertsTable.js
@@ -41,7 +41,6 @@ const AlertsTable = ({
   handleRefreshWithFilters,
   isCrossProjects,
   pageData,
-  paginationConfigAlertsRef,
   requestErrorMessage,
   selectedAlert,
   tableContent,
@@ -115,17 +114,12 @@ AlertsTable.propTypes = {
   filters: PropTypes.object.isRequired,
   filtersStore: PropTypes.object.isRequired,
   handleCancel: PropTypes.func,
-  handleRefreshAlerts: PropTypes.func.isRequired,
   handleRefreshWithFilters: PropTypes.func.isRequired,
   isAlertsPage: PropTypes.bool,
   isCrossProjects: PropTypes.bool.isRequired,
-  navigate: PropTypes.func,
-  navigateToPerProjectAlertsPage: PropTypes.func,
   pageData: PropTypes.object.isRequired,
-  paginationConfigAlertsRef: PropTypes.object.isRequired,
   requestErrorMessage: PropTypes.string.isRequired,
   selectedAlert: PropTypes.object.isRequired,
-  setSearchParams: PropTypes.func.isRequired,
   tableContent: PropTypes.arrayOf(PropTypes.object).isRequired,
   toggleRow: PropTypes.func
 }
