+++ src/elements/AlertsTableRow/AlertsTableRow.scss
@@ -2,8 +2,6 @@
 @import '~igz-controls/scss/borders';
 
 .alert-row {
-  flex-wrap: wrap;
-
   .table-body__cell.alert-row-notification-cell {
     min-width: 150px;
   }
