+++ src/hooks/useAlertsPageData.js
@@ -17,78 +17,20 @@ illegal under applicable law, and the grant of the foregoing license
 under the Apache 2.0 license is conditioned upon your compliance with
 such restriction.
 */
-import { useCallback, useRef, useState } from 'react'
-import { useParams } from 'react-router-dom'
-import { useDispatch } from 'react-redux'
-
-import {
-  BE_PAGE,
-  BE_PAGE_SIZE,
-  MODEL_ENDPOINT_ID,
-  PROJECT_FILTER,
-  PROJECTS_FILTER_ALL_ITEMS
-} from '../constants'
+import { useRefreshAlerts } from './useRefreshAlerts.hook'
 import { usePagination } from './usePagination.hook'
-import { fetchAlerts } from '../reducers/alertsReducer'
 
 export const useAlertsPageData = (filters, isAlertsPage) => {
-  const [alerts, setAlerts] = useState(null)
-  const [requestErrorMessage, setRequestErrorMessage] = useState('')
-  const paginationConfigAlertsRef = useRef({})
-  const lastCheckedAlertIdRef = useRef(null)
-
-  const abortControllerRef = useRef(new AbortController())
-
-  const params = useParams()
-  const dispatch = useDispatch()
-
-  const refreshAlerts = useCallback(
-    filters => {
-      setAlerts(null)
-      lastCheckedAlertIdRef.current = null
-      abortControllerRef.current = new AbortController()
-      const projectName = !isAlertsPage
-        ? params.projectName || params.id
-        : filters?.[PROJECT_FILTER]?.toLowerCase?.() !== PROJECTS_FILTER_ALL_ITEMS &&
-            params?.projectName !== PROJECTS_FILTER_ALL_ITEMS
-          ? filters[PROJECT_FILTER]?.toLowerCase()
-          : params.id || params.projectName
-
-      if (!isAlertsPage) {
-        filters = { ...filters, [MODEL_ENDPOINT_ID]: params.tag }
-      }
-
-      dispatch(
-        fetchAlerts({
-          project: projectName,
-          filters,
-          config: {
-            ui: {
-              controller: abortControllerRef.current,
-              setRequestErrorMessage
-            },
-            params: {
-              page: paginationConfigAlertsRef.current[BE_PAGE],
-              'page-size': paginationConfigAlertsRef.current[BE_PAGE_SIZE]
-            }
-          }
-        })
-      )
-        .unwrap()
-        .then(response => {
-          if (response?.activations) {
-            setAlerts(response.activations.length > 0 ? response.activations : [])
-            paginationConfigAlertsRef.current.paginationResponse = response.pagination
-          } else {
-            setAlerts([])
-          }
-        })
-        .catch(() => {
-          setAlerts([])
-        })
-    },
-    [dispatch, isAlertsPage, params.id, params.projectName, params.tag]
-  )
+  const {
+    abortControllerRef,
+    alerts,
+    lastCheckedAlertIdRef,
+    paginationConfigAlertsRef,
+    refreshAlerts,
+    requestErrorMessage,
+    setAlerts,
+    params
+  } = useRefreshAlerts(filters, isAlertsPage)
 
   const [handleRefreshAlerts, paginatedAlerts, searchParams, setSearchParams] = usePagination({
     content: alerts ?? [],
